feat(@dpc-sdp/ripple-tide-search): :sparkles: add options from taxonomy support in custom col
fixes [R20-1692]
