+++ packages/ripple-tide-search/mapping/components/custom-collection/custom-collection-mapping.ts
@@ -1,14 +1,21 @@
 import { TideDynamicPageComponent } from '@dpc-sdp/ripple-tide-api/types'
-
-export const customCollectionMapping = (
-  field
+import { processConfig } from './../../utils'
+export const customCollectionMapping = async (
+  field,
+  page,
+  tidePageApi
 ): TideDynamicPageComponent<any> => {
+  const config = await processConfig(
+    field.field_content_collection_config,
+    tidePageApi
+  )
+
   return {
     component: 'TideCustomCollection',
     id: field.drupal_internal__id.toString(),
     props: {
       id: field.drupal_internal__id.toString(),
-      ...field.field_content_collection_config
+      ...config
     }
   }
 }
